refactor: move intent endpoint back into search path
